backend/requirements.txt
fastapi
uvicorn
pydantic
twilio
python-dotenv
supabase
